# springboot/swagger2doc-master/swagger2doc-master/resources/swagger/swagger.json
{"swagger":"2.0","info":{"description":"Restful API","version":"1.0","title":"test Restful  api","termsOfService":"http://abocode.com/"},"host":"127.0.0.1:8089","basePath":"/","tags":[{"name":"Word API","description":"Word Controller"},{"name":"配置信息","description":"Config Controller"}],"paths":{"/configs":{"post":{"tags":["配置信息"],"summary":"添加配置信息","description":"添加配置信息<br>错误码说明：<br/>BASIC_1000:该城市已经存在<br>","operationId":"createUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"command","description":"command","required":true,"schema":{"$ref":"#/definitions/配置信息"}}],"responses":{"201":{"description":"Created"},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/configs/code/{code}":{"get":{"tags":["配置信息"],"summary":"根据code获得配置信息","description":"根据code获得配置信息","operationId":"getByNameUsingGET","produces":["*/*"],"parameters":[{"name":"code","in":"path","description":"code","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"type":"string"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/configs/{id}":{"get":{"tags":["配置信息"],"summary":"根据配置信息id获取配置信息","description":"根据配置信息id获取配置信息","operationId":"getUsingGET","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/配置信息"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false},"put":{"tags":["配置信息"],"summary":"修改配置信息","description":"修改配置信息","operationId":"updateUsingPUT","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"command","description":"command","required":true,"schema":{"$ref":"#/definitions/配置信息"}},{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int32"}],"responses":{"201":{"description":"Created"},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/fileToWord":{"post":{"tags":["Word API"],"summary":"将 swagger json文件转换成 word文档并下载","operationId":"getWordUsingPOST","consumes":["multipart/form-data"],"produces":["*/*"],"parameters":[{"name":"jsonFile","in":"formData","description":"swagger json file","required":false,"type":"file"}],"responses":{"200":{"description":"请求成功。"},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/strToWord":{"post":{"tags":["Word API"],"summary":"将 swagger json字符串转换成 word文档并下载","operationId":"getWordUsingPOST_1","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"jsonStr","description":"swagger json string","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"请求成功。"},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/urlToWord":{"get":{"tags":["Word API"],"summary":"将 swagger json url地址转换成 word文档并下载","operationId":"wordUsingGET","produces":["*/*"],"parameters":[{"name":"url","in":"query","description":"资源地址","required":false,"type":"string","allowEmptyValue":false}],"responses":{"200":{"description":"请求成功。"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}}},"definitions":{"File":{"type":"object","properties":{"absolute":{"type":"boolean"},"absoluteFile":{"$ref":"#/definitions/File"},"absolutePath":{"type":"string"},"canonicalFile":{"$ref":"#/definitions/File"},"canonicalPath":{"type":"string"},"directory":{"type":"boolean"},"executable":{"type":"boolean"},"file":{"type":"boolean"},"freeSpace":{"type":"integer","format":"int64"},"hidden":{"type":"boolean"},"lastModified":{"type":"integer","format":"int64"},"name":{"type":"string"},"parent":{"type":"string"},"parentFile":{"$ref":"#/definitions/File"},"path":{"type":"string"},"readable":{"type":"boolean"},"totalSpace":{"type":"integer","format":"int64"},"usableSpace":{"type":"integer","format":"int64"},"writable":{"type":"boolean"}},"title":"File"},"InputStream":{"type":"object","title":"InputStream"},"Resource":{"type":"object","properties":{"description":{"type":"string"},"file":{"$ref":"#/definitions/File"},"filename":{"type":"string"},"inputStream":{"$ref":"#/definitions/InputStream"},"open":{"type":"boolean"},"readable":{"type":"boolean"},"uri":{"$ref":"#/definitions/URI"},"url":{"$ref":"#/definitions/URL"}},"title":"Resource"},"URI":{"type":"object","properties":{"absolute":{"type":"boolean"},"authority":{"type":"string"},"fragment":{"type":"string"},"host":{"type":"string"},"opaque":{"type":"boolean"},"path":{"type":"string"},"port":{"type":"integer","format":"int32"},"query":{"type":"string"},"rawAuthority":{"type":"string"},"rawFragment":{"type":"string"},"rawPath":{"type":"string"},"rawQuery":{"type":"string"},"rawSchemeSpecificPart":{"type":"string"},"rawUserInfo":{"type":"string"},"scheme":{"type":"string"},"schemeSpecificPart":{"type":"string"},"userInfo":{"type":"string"}},"title":"URI"},"URL":{"type":"object","properties":{"authority":{"type":"string"},"content":{"type":"object"},"defaultPort":{"type":"integer","format":"int32"},"deserializedFields":{"$ref":"#/definitions/URLStreamHandler"},"file":{"type":"string"},"host":{"type":"string"},"path":{"type":"string"},"port":{"type":"integer","format":"int32"},"protocol":{"type":"string"},"query":{"type":"string"},"ref":{"type":"string"},"serializedHashCode":{"type":"integer","format":"int32"},"userInfo":{"type":"string"}},"title":"URL"},"URLStreamHandler":{"type":"object","title":"URLStreamHandler"},"配置信息":{"type":"object","required":["code","description","name","value"],"properties":{"code":{"type":"string","description":"代码"},"description":{"type":"string","description":"描述"},"name":{"type":"string","description":"名称,2-20"},"unit":{"type":"string","description":"单位"},"value":{"type":"string","description":"值"}},"title":"配置信息"}}}